+++ tensorflow_datasets/image/mnist.py
@@ -317,11 +317,7 @@ class EMNIST(MNIST):
             "The EMNIST dataset is a set of handwritten character digits "
             "derived from the NIST Special Database 19 and converted to "
             "a 28x28 pixel image format and dataset structure that directly "
-            "matches the MNIST dataset.\n\n"
-            "Note: Like the original EMNIST data, images provided here are "
-            "inverted horizontally and rotated 90 anti-clockwise. You can use "
-            "`tf.transpose` within `ds.map` to convert the images to a "
-            "human-friendlier format."),
+            "matches the MNIST dataset."),
         features=tfds.features.FeaturesDict({
             "image":
                 tfds.features.Image(shape=MNIST_IMAGE_SHAPE),