+++ tensorflow_datasets/text/super_glue_test.py
@@ -29,19 +29,10 @@ _BASE_DIR = os.path.join(
     "testing/test_data/fake_examples/super_glue")
 
 
-class SuperGlueBoolQTest(testing.DatasetBuilderTestCase):
-  BUILDER_CONFIG_NAMES_TO_TEST = ["boolq"]
-  DATASET_CLASS = super_glue.SuperGlue
-  SPLITS = {
-      "train": 3,
-      "validation": 2,
-      "test": 1,
-  }
-
-
 class SuperGlueCbTest(testing.DatasetBuilderTestCase):
   BUILDER_CONFIG_NAMES_TO_TEST = ["cb"]
   DATASET_CLASS = super_glue.SuperGlue
+  EXAMPLE_DIR = os.path.join(_BASE_DIR, "CB")
   SPLITS = {
       "train": 3,
       "validation": 2,
@@ -52,6 +43,7 @@ class SuperGlueCbTest(testing.DatasetBuilderTestCase):
 class SuperGlueCopaTest(testing.DatasetBuilderTestCase):
   BUILDER_CONFIG_NAMES_TO_TEST = ["copa"]
   DATASET_CLASS = super_glue.SuperGlue
+  EXAMPLE_DIR = os.path.join(_BASE_DIR, "COPA")
   SPLITS = {
       "train": 3,
       "validation": 2,
@@ -62,6 +54,7 @@ class SuperGlueCopaTest(testing.DatasetBuilderTestCase):
 class SuperGlueMultiRcTest(testing.DatasetBuilderTestCase):
   BUILDER_CONFIG_NAMES_TO_TEST = ["multirc"]
   DATASET_CLASS = super_glue.SuperGlue
+  EXAMPLE_DIR = os.path.join(_BASE_DIR, "MultiRC")
   SPLITS = {
       "train": 58,
       "validation": 62,
@@ -69,19 +62,10 @@ class SuperGlueMultiRcTest(testing.DatasetBuilderTestCase):
   }
 
 
-class SuperGlueReCoRDTest(testing.DatasetBuilderTestCase):
-  BUILDER_CONFIG_NAMES_TO_TEST = ["record"]
-  DATASET_CLASS = super_glue.SuperGlue
-  SPLITS = {
-      "train": 9,
-      "validation": 2,
-      "test": 1,
-  }
-
-
 class SuperGlueRteTest(testing.DatasetBuilderTestCase):
   BUILDER_CONFIG_NAMES_TO_TEST = ["rte"]
   DATASET_CLASS = super_glue.SuperGlue
+  EXAMPLE_DIR = os.path.join(_BASE_DIR, "RTE")
   SPLITS = {
       "train": 3,
       "validation": 2,
@@ -92,6 +76,7 @@ class SuperGlueRteTest(testing.DatasetBuilderTestCase):
 class SuperGlueWscTest(testing.DatasetBuilderTestCase):
   BUILDER_CONFIG_NAMES_TO_TEST = ["wsc", "wsc.fixed"]
   DATASET_CLASS = super_glue.SuperGlue
+  EXAMPLE_DIR = os.path.join(_BASE_DIR, "WSC")
   SPLITS = {
       "train": 3,
       "validation": 2,
@@ -102,12 +87,38 @@ class SuperGlueWscTest(testing.DatasetBuilderTestCase):
 class SuperGlueWicTest(testing.DatasetBuilderTestCase):
   BUILDER_CONFIG_NAMES_TO_TEST = ["wic"]
   DATASET_CLASS = super_glue.SuperGlue
+  EXAMPLE_DIR = os.path.join(_BASE_DIR, "WiC")
   SPLITS = {
       "train": 3,
       "validation": 2,
       "test": 1,
   }
 
+
+class SuperGlueCbS3Test(SuperGlueCbTest):
+  VERSION = "experimental_latest"
+
+
+class SuperGlueCopaS3Test(SuperGlueCopaTest):
+  VERSION = "experimental_latest"
+
+
+class SuperGlueMultiRcS3Test(SuperGlueMultiRcTest):
+  VERSION = "experimental_latest"
+
+
+class SuperGlueRteS3Test(SuperGlueRteTest):
+  VERSION = "experimental_latest"
+
+
+class SuperGlueWscS3Test(SuperGlueWscTest):
+  VERSION = "experimental_latest"
+
+
+class SuperGlueWicS3Test(SuperGlueWicTest):
+  VERSION = "experimental_latest"
+
+# TODO(adarob): Add MultiRC test.
 
 if __name__ == "__main__":
   testing.test_main()